Private module changes
